Split out example tests into their testing module so this is clearer.
